Locale :: init moved of analyse
