+++ src/Backend/Modules/Locale/Actions/Analyse.php
@@ -14,7 +14,7 @@ use Backend\Core\Engine\Authentication as BackendAuthentication;
 use Backend\Core\Engine\DataGridArray as BackendDataGridArray;
 use Backend\Core\Engine\Language as BL;
 use Backend\Core\Engine\Model as BackendModel;
-use Backend\Modules\Locale\Engine\Model as BackendLocaleModel;
+use Backend\Modules\Locale\Engine\AnalyseModel as BackendLocaleModel;
 
 /**
  * This is the analyse-action, it will display an overview of used locale.
@@ -76,7 +76,8 @@ class Analyse extends BackendBaseActionIndex
         /*
          * Frontend datagrid
          */
-        $this->dgFrontend = new BackendDataGridArray(BackendLocaleModel::getNonExistingFrontendLocale(BL::getWorkingLanguage()));
+        $getNonExistingFrontendLocale = BackendLocaleModel::getNonExistingFrontendLocale(BL::getWorkingLanguage());
+        $this->dgFrontend = new BackendDataGridArray($getNonExistingFrontendLocale);
 
         // overrule default URL
         $this->dgFrontend->setURL(BackendModel::createURLForAction(null, null, null, array('offset' => '[offset]', 'order' => '[order]', 'sort' => '[sort]'), false));
@@ -112,7 +113,8 @@ class Analyse extends BackendBaseActionIndex
         /*
          * Backend datagrid
          */
-        $this->dgBackend = new BackendDataGridArray(BackendLocaleModel::getNonExistingBackendLocale(BL::getWorkingLanguage()));
+        $getNonExistingBackendLocale = BackendLocaleModel::getNonExistingBackendLocale(BL::getWorkingLanguage());
+        $this->dgBackend = new BackendDataGridArray($getNonExistingBackendLocale);
 
         // overrule default URL
         $this->dgBackend->setURL(BackendModel::createURLForAction(null, null, null, array('offset' => '[offset]', 'order' => '[order]', 'sort' => '[sort]'), false));
